catch-all/06_bots_telegram/10_mareas_bot/requirements.txt
aiohttp==3.10.9
beautifulsoup4==4.12.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-telegram-bot>=20.0
redis==5.2.1
requests==2.31.0